Se realizan cambios para corregir errores al construir en windows On branch master Changes to be committed: 	modified:   CMakeLists.txt 	modified:   appdb/CMakeLists.txt 	modified:   cpuinfo/include/cpuinfo.h 	new file:   cpuinfo/src/cpuinfo_win.cpp 	new file:   fileinfo/src/fileinfo_win.cpp 	modified:   socket/src/socket_win32.cpp

+++ cpuinfo/include/cpuinfo.h
@@ -7,7 +7,7 @@ using std::string;
 
 class CPUInfo {
 	public:
-		long frequency;
+		long long frequency;
 		string name;
 		int cores;
 		long cache_size;
